Move impls for MapType and MapTypeRef
