+++ crates/cgp-core/src/prelude.rs
@@ -6,12 +6,12 @@ pub use cgp_component::{
     UseFields, WithContext, WithProvider,
 };
 pub use cgp_error::{CanRaiseError, CanWrapError, HasErrorType};
-pub use cgp_field::impls::UseField;
+pub use cgp_field::impls::{IsMut, IsNothing, IsPresent, IsRef, IsVoid, UseField};
 pub use cgp_field::traits::{
     BuildField, ExtractField, FieldGetter, FinalizeBuild, FinalizeExtract, FromFields, FromVariant,
     HasBuilder, HasExtractor, HasExtractorMut, HasExtractorRef, HasField, HasFieldMut, HasFields,
-    HasFieldsRef, IntoBuilder, IsMut, IsNothing, IsPresent, IsRef, IsVoid, MapType, MapTypeRef,
-    MutFieldGetter, PartialData, ToFields, ToFieldsRef, UpdateField,
+    HasFieldsRef, IntoBuilder, MapType, MapTypeRef, MutFieldGetter, PartialData, ToFields,
+    ToFieldsRef, UpdateField,
 };
 pub use cgp_field::types::{Char, Cons, Either, Field, Index, Nil, Void};
 pub use cgp_macro::{
