+++ crates/cgp-field/src/impls/build_default.rs
@@ -1,7 +1,5 @@
-use crate::impls::CanBuildFrom;
-use crate::traits::{
-    FinalizeBuild, HasBuilder, IsNothing, IsPresent, TransformMap, TransformMapFields,
-};
+use crate::impls::{CanBuildFrom, IsNothing, IsPresent};
+use crate::traits::{FinalizeBuild, HasBuilder, TransformMap, TransformMapFields};
 
 pub trait CanBuildWithDefault<Source> {
     fn build_with_default(source: Source) -> Self;
