+++ crates/cgp-field/src/impls/mod.rs
@@ -2,6 +2,8 @@ mod build_default;
 mod build_from;
 mod cast;
 mod chain;
+mod map_type;
+mod map_type_ref;
 mod use_field;
 mod use_ref;
 
@@ -9,5 +11,7 @@ pub use build_default::*;
 pub use build_from::*;
 pub use cast::*;
 pub use chain::*;
+pub use map_type::*;
+pub use map_type_ref::*;
 pub use use_field::*;
 pub use use_ref::*;
