+++ crates/cgp-field/src/traits/map_type.rs
@@ -1,29 +1,3 @@
-use crate::types::Void;
-
 pub trait MapType {
     type Map<T>;
-}
-
-pub struct IsPresent;
-
-impl MapType for IsPresent {
-    type Map<T> = T;
-}
-
-pub struct IsNothing;
-
-impl MapType for IsNothing {
-    type Map<T> = ();
-}
-
-pub struct IsVoid;
-
-impl MapType for IsVoid {
-    type Map<T> = Void;
-}
-
-pub struct IsOptional;
-
-impl MapType for IsOptional {
-    type Map<T> = Option<T>;
 }
