+++ crates/cgp-field/src/traits/map_type_ref.rs
@@ -1,21 +1,3 @@
 pub trait MapTypeRef {
     type Map<'a, T: 'a>: 'a;
 }
-
-pub struct IsRef;
-
-impl MapTypeRef for IsRef {
-    type Map<'a, T: 'a> = &'a T;
-}
-
-pub struct IsMut;
-
-impl MapTypeRef for IsMut {
-    type Map<'a, T: 'a> = &'a mut T;
-}
-
-pub struct IsOwned;
-
-impl MapTypeRef for IsOwned {
-    type Map<'a, T: 'a> = T;
-}
